+++ apps/web/public/_routes.json
@@ -1,0 +1,5 @@
+{
+	"version": 1,
+	"include": ["/*"],
+	"exclude": ["/blog/*"]
+}
